fix: cl ..f

+++ src/cl/mandelbrot.cl
@@ -38,7 +38,7 @@ __kernel void mandelbrot(__global float* results,
     }
     float result = iter;
     if (smoothing && iter != iters) {
-        result = result - logf(logf(sqrtf(x * x + y * y)) / logf(threshold)) / logf(2.0f);
+        result = result - log(log(sqrt(x * x + y * y)) / log(threshold)) / log(2.0f);
     }
     result = 1.0f * result / iters;
     results[j * width + i] = result;
